New convention for asset folders when they are not at Workshop yet Fixes at TextureUtils

+++ Redirectors/UIWorkshopAssetRedirector.cs
@@ -1,5 +1,4 @@
 ﻿using ColossalFramework.Packaging;
-using ColossalFramework.UI;
 using Klyte.Commons.Extensions;
 using Klyte.Commons.Utils;
 using System.IO;
@@ -22,40 +21,47 @@ namespace Klyte.Commons.Redirectors
         {
             var m_ContentPath = __instance.GetType().GetField("m_ContentPath", RedirectorUtils.allFlags).GetValue(__instance) as string;
             var m_TargetAsset = __instance.GetType().GetField("m_TargetAsset", RedirectorUtils.allFlags).GetValue(__instance) as Package.Asset;
-            var rootAssetFolder = Path.GetDirectoryName(m_TargetAsset.package.packagePath);
-            LogUtils.DoErrorLog($"rootAssetFolder2: {rootAssetFolder}; ");
-            bool bundledAnyFile = false;
-            if (!(CommonProperties.AssetExtraFileNames is null))
+            if (m_TargetAsset.isMainAsset)
             {
-                foreach (string filename in CommonProperties.AssetExtraFileNames)
+                var rootAssetFolder = FileUtils.GetRootFolderForK45(m_TargetAsset.package);
+                LogUtils.DoErrorLog($"rootAssetFolder: {rootAssetFolder}; ");
+                if (!Directory.Exists(rootAssetFolder))
                 {
-                    var targetFilename = Path.Combine(rootAssetFolder, filename);
-                    if (File.Exists(targetFilename))
+                    return;
+                }
+
+                bool bundledAnyFile = false;
+                if (!(CommonProperties.AssetExtraFileNames is null))
+                {
+                    foreach (string filename in CommonProperties.AssetExtraFileNames)
                     {
-                        File.Copy(targetFilename, Path.Combine(m_ContentPath, filename));
-                        bundledAnyFile = true;
+                        var targetFilename = Path.Combine(rootAssetFolder, filename);
+                        if (File.Exists(targetFilename))
+                        {
+                            File.Copy(targetFilename, Path.Combine(m_ContentPath, filename));
+                            bundledAnyFile = true;
+                        }
                     }
                 }
-            }
-            if (!(CommonProperties.AssetExtraDirectoryNames is null))
-            {
-                foreach (string directory in CommonProperties.AssetExtraDirectoryNames)
+                if (!(CommonProperties.AssetExtraDirectoryNames is null))
                 {
-                    var targetFolder = Path.Combine(rootAssetFolder, directory);
-                    if (Directory.Exists(targetFolder))
+                    foreach (string directory in CommonProperties.AssetExtraDirectoryNames)
                     {
-                        WorkshopHelper.DirectoryCopy(targetFolder, Path.Combine(m_ContentPath, directory), true);
-                        bundledAnyFile = true;
+                        var targetFolder = Path.Combine(rootAssetFolder, directory);
+                        if (Directory.Exists(targetFolder))
+                        {
+                            WorkshopHelper.DirectoryCopy(targetFolder, Path.Combine(m_ContentPath, directory), true);
+                            bundledAnyFile = true;
+                        }
                     }
                 }
-            }
 
-            if (bundledAnyFile)
-            {
-                var tagsField = (__instance.GetType().GetField("m_Tags", RedirectorUtils.allFlags));
-                tagsField.SetValue(__instance, (tagsField.GetValue(__instance) as string[]).Concat(new string[] { CommonProperties.ModName, $"K45 {CommonProperties.Acronym}" }).Distinct().ToArray());
+                if (bundledAnyFile)
+                {
+                    var tagsField = (__instance.GetType().GetField("m_Tags", RedirectorUtils.allFlags));
+                    tagsField.SetValue(__instance, (tagsField.GetValue(__instance) as string[]).Concat(new string[] { CommonProperties.ModName, $"K45 {CommonProperties.Acronym}" }).Distinct().ToArray());
+                }
             }
-
         }
     }
 }
